Revert "Implements more packets"
This reverts commit 57498a611de6cc8f21f77154aa27fe9cebd8592c.

+++ src/main/java/cn/nukkit/inventory/transaction/CraftingTransaction.java
@@ -106,7 +106,7 @@ public class CraftingTransaction extends InventoryTransaction {
 		 * transaction goes wrong.
 		 */
         ContainerClosePacket pk = new ContainerClosePacket();
-        pk.windowId = (byte) ContainerIds.NONE;
+        pk.windowId = ContainerIds.NONE;
         source.getServer().getScheduler().scheduleDelayedTask(new Task() {
             @Override
             public void onRun(int currentTick) {
